Move topics state related files
