--- src/modules/topics/injection/index.js
@@ -1,6 +1,0 @@
-import { TopicsPlugin } from '../../topics/store/TopicsPlugin';
-
-export const topicsModule = ($injector) => {
-	$injector
-		.registerSingleton('TopicsPlugin', new TopicsPlugin());
-};
